# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "i9lyfe Social Media Backend API.",
        "title": "i9lyfe Backend API",
        "contact": {
            "name": "i9ine",
            "email": "oluwarinolasam@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
        "/app/private/me": {
            "get": {
                "description": "Get info on the user currently in session",
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "app/private"
                ],
                "summary": "Get session user",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "User session request cookie",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User info",
                        "schema": {
                            "$ref": "#/definitions/UITypes.ClientUser"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/app/private/me/change_profile_picture": {
            "put": {
                "description": "Change user profile picture",
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "app/private"
                ],
                "summary": "Change user profile picture",
                "parameters": [
                    {
                        "description": "Profile picture data",
                        "name": "picture_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "User session request cookie",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Done",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/app/private/me/edit_profile": {
            "put": {
                "description": "Edit user profile",
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "app/private"
                ],
                "summary": "Edit user profile",
                "parameters": [
                    {
                        "description": "User's Name field",
                        "name": "name",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "User's Birthday field in milliseconds since Unix Epoch",
                        "name": "birthday",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "User's Bio field",
                        "name": "bio",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "User session request cookie",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Done",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/app/private/me/signout": {
            "get": {
                "description": "Signout the user currently in session",
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "app/private"
                ],
                "summary": "Signout user",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "User session request cookie",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bye message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/app/private/users/:username/follow": {
            "post": {
                "description": "Follow user",
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "app/private"
                ],
                "summary": "Follow user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User to follow",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "User session request cookie",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Done",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Validation error | User trying to follow self",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/app/private/users/:username/unfollow": {
            "delete": {
                "description": "Unfollow user",
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "app/private"
                ],
                "summary": "Unfollow user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User to unfollow",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "User session request cookie",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Done",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "UITypes.ClientUser": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "profilePicUrl": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "appErrors.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT API key in encrypted cookie to protect private endpoints",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}